Refactor two scan operations to validate the outputs with the host sequential operation.
