+++ 07_parallel_programming_pattern/03_scan/Makefile
@@ -4,22 +4,29 @@ NVCC=${CUDA_PATH}/bin/nvcc -ccbin ${HOST_COMPILER}
 TARGET=scan
 
 INCLUDES = -I${CUDA_PATH}/samples/common/inc
-NVCC_FLAGS=-m64 -G # --resource-usage
+NVCC_FLAGS=-m64 --resource-usage -lineinfo
 
 # Gencode argumentes
-SMS = 30 35 37 50 52 60 61 70
+SMS = 30 35 37 50 52 60 61 70 75
 $(foreach sm, ${SMS}, $(eval GENCODE_FLAGS += -gencode arch=compute_$(sm),code=sm_$(sm)))
 
 LIBRARIES += -L/usr/local/cuda/lib
-ALL_CCFLAGS += -std=c++11 $(INCLUDES) $(LIBRARIES)
+ALL_CCFLAGS += -std=c++11 $(NVCC_FLAGS) $(INCLUDES) $(LIBRARIES)
 
 all : ${TARGET}
 
-scan: scan.cu
+scan_v1.o: scan_v1.cu
+	$(EXEC) $(NVCC) $(INCLUDES) $(ALL_CCFLAGS) $(GENCODE_FLAGS) -o $@ -c $<
+
+scan_v2.o: scan_v2.cu
+	$(EXEC) $(NVCC) $(INCLUDES) $(ALL_CCFLAGS) $(GENCODE_FLAGS) -o $@ -c $<
+
+scan: scan.cu scan_v1.o scan_v2.o
 	$(EXEC) $(NVCC) $(ALL_CCFLAGS) $(GENCODE_FLAGS) -o $@ $+
 
 nvprof: scan
-	nvprof -f -o $+.nvvp ./$+
+	nvprof -f -o $+.nvvp --profile-from-start off ./$+
+	nvprof -f -o $+-metrics.nvvp --analysis-metrics ./$+
 
 clean:
-	rm -f ${TARGET} *.o
+	rm -f ${TARGET} *.o *.nvvp
