+++ 07_parallel_programming_pattern/03_scan/scan.h
@@ -1,41 +1,12 @@
-#include <cuda_runtime.h>
-#include <stdarg.h>
+#ifndef _SCAN_H_
+#define _SCAN_H_
 
-#ifndef _UTILS_H_
-#define _UTILS_H_
+#define BLOCK_DIM 512
 
-#define NUM_ITEM 16
-#define BLOCK_DIM 8
+#define DEBUG_INDEX         0
+#define DEBUG_OUTPUT_NUM    16
 
-#define DEBUG_OUTPUT_NUM 16
+void scan_v1(float *d_output, float *d_input, int length);
+void scan_v2(float *d_output, float *d_input, int length);
 
-// generate input data
-void generate_data(float *ptr, int length)
-{
-    // fill the buffer with random generated unsigned integers
-    for (int i = 0; i < length; i++)
-        ptr[i] = float(rand() - RAND_MAX / 2) / RAND_MAX;
-}
-
-bool validation(float *a, float *b, int length)
-{
-    float epsilon = 0.000001;
-    bool result = true;
-    for (int i = 0; i < length; i++)
-        if (abs(a[i] - b[i]) >= epsilon)
-            result = false;
-    return result;
-}
-
-void print_val(float *h_list, int length, ...)
-{
-    va_list argptr;
-    va_start(argptr, length);
-
-    printf("%s\t", va_arg(argptr, char *));
-    for (int i = 0; i < length; i++)
-        printf("%.3f\t", h_list[i]);
-    printf("\n");
-}
-
-#endif  // _UTILS_H_
+#endif // _SCAN_H_
