+++ 07_parallel_programming_pattern/03_scan/scan_v1.cu
@@ -1,24 +1,21 @@
 #include "scan.h"
 
 __global__ void
-scan_v1_kernel(float *d_output, float *d_input, int length, int offset)
+scan_v1_kernel(float *d_output, float *d_input, int length)
 {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
-
-    if (idx >= length)
-        return;
-
-    if (idx - offset >= 0)
-        d_output[idx] += d_input[idx - offset];
+        
+    float element = 0.f;
+    for (int offset = 0; offset < length; offset++) {
+        if (idx - offset >= 0)
+            element += d_input[idx - offset];
+    }
+    d_output[idx] = element;
 }
 
 void scan_v1(float *d_output, float *d_input, int length)
 {
     dim3 dimBlock(BLOCK_DIM);
     dim3 dimGrid((length + BLOCK_DIM - 1) / BLOCK_DIM);
-    for (int offset = 0; offset < length; offset++)
-    {
-        scan_v1_kernel<<<dimGrid, dimBlock>>>(d_output, d_input, length, offset);
-        cudaDeviceSynchronize();
-    }
+    scan_v1_kernel<<<dimGrid, dimBlock>>>(d_output, d_input, length);
 }
